+++ wcfsetup/install/files/js/WCF.Message.js
@@ -2427,7 +2427,7 @@ WCF.Message.Quote.Manager = Class.extend({
 					}
 				}
 				
-				if (insert) this._editorElement.redactor('wbbcode.insertQuoteBBCode', $username, $link, $quote, $quote);
+				if (insert && this._editorElement.length) this._editorElement.redactor('wbbcode.insertQuoteBBCode', $username, $link, $quote, $quote);
 			}
 			else {
 				this._editorElementAlternative.redactor('wbbcode.insertQuoteBBCode', $username, $link, $quote, $quote);