+++ wcfsetup/install/files/js/WoltLab/WCF/Ajax/Request.js
@@ -35,6 +35,7 @@ define(['Core', 'Language', 'Dom/ChangeListener', 'Dom/Util', 'Ui/Dialog', 'Wolt
 			this._options = Core.extend({
 				// request data
 				data: {},
+				contentType: 'application/x-www-form-urlencoded; charset=UTF-8',
 				responseType: 'application/json',
 				type: 'POST',
 				url: '',
@@ -49,6 +50,8 @@ define(['Core', 'Language', 'Dom/ChangeListener', 'Dom/Util', 'Ui/Dialog', 'Wolt
 				failure: null,
 				finalize: null,
 				success: null,
+				progress: null,
+				uploadProgress: null,
 				
 				callbackObject: null
 			}, options);
@@ -67,6 +70,8 @@ define(['Core', 'Language', 'Dom/ChangeListener', 'Dom/Util', 'Ui/Dialog', 'Wolt
 				if (typeof this._options.callbackObject._ajaxFailure === 'function') this._options.failure = this._options.callbackObject._ajaxFailure.bind(this._options.callbackObject);
 				if (typeof this._options.callbackObject._ajaxFinalize === 'function') this._options.finalize = this._options.callbackObject._ajaxFinalize.bind(this._options.callbackObject);
 				if (typeof this._options.callbackObject._ajaxSuccess === 'function') this._options.success = this._options.callbackObject._ajaxSuccess.bind(this._options.callbackObject);
+				if (typeof this._options.callbackObject._ajaxProgress === 'function') this._options.progress = this._options.callbackObject._ajaxProgress.bind(this._options.callbackObject);
+				if (typeof this._options.callbackObject._ajaxUploadProgress === 'function') this._options.uploadProgress = this._options.callbackObject._ajaxUploadProgress.bind(this._options.callbackObject);
 			}
 			
 			if (_didInit === false) {
@@ -96,7 +101,9 @@ define(['Core', 'Language', 'Dom/ChangeListener', 'Dom/Util', 'Ui/Dialog', 'Wolt
 			
 			this._xhr = new XMLHttpRequest();
 			this._xhr.open(this._options.type, this._options.url, true);
-			this._xhr.setRequestHeader('Content-Type', 'application/x-www-form-urlencoded; charset=UTF-8');
+			if (this._options.contentType) {
+				this._xhr.setRequestHeader('Content-Type', this._options.contentType);
+			}
 			this._xhr.setRequestHeader('X-Requested-With', 'XMLHttpRequest');
 			
 			var self = this;
@@ -121,9 +128,16 @@ define(['Core', 'Language', 'Dom/ChangeListener', 'Dom/Util', 'Ui/Dialog', 'Wolt
 				self._failure(this, options);
 			};
 			
+			if (this._options.progress) {
+				this._xhr.onprogress = this._options.progress;
+			}
+			if (this._options.uploadProgress) {
+				this._xhr.upload.onprogress = this._options.uploadProgress;
+			}
+			
 			if (this._options.type === 'POST') {
 				var data = this._options.data;
-				if (typeof data === 'object') {
+				if (typeof data === 'object' && Core.getType(data) !== 'FormData') {
 					data = Core.serialize(data);
 				}
 				
@@ -180,7 +194,7 @@ define(['Core', 'Language', 'Dom/ChangeListener', 'Dom/Util', 'Ui/Dialog', 'Wolt
 		 * @param	{object<string, *>}	data	request data
 		 */
 		setData: function(data) {
-			if (this._data !== null) {
+			if (this._data !== null && Core.getType(data) !== 'FormData') {
 				data = Core.extend(this._data, data);
 			}
 			
@@ -270,6 +284,7 @@ define(['Core', 'Language', 'Dom/ChangeListener', 'Dom/Util', 'Ui/Dialog', 'Wolt
 				
 				var html = '<div class="ajaxDebugMessage"><p>' + message + '</p>' + details + '</div>';
 				
+				if (UiDialog === undefined) UiDialog = require('Ui/Dialog');
 				UiDialog.openStatic(DomUtil.getUniqueId(), html, {
 					title: Language.get('wcf.global.error.title')
 				});