+++ wcfsetup/install/files/lib/data/sitemap/Sitemap.class.php
@@ -5,7 +5,6 @@ use wcf\data\TDatabaseObjectOptions;
 use wcf\data\TDatabaseObjectPermissions;
 use wcf\system\exception\SystemException;
 use wcf\system\WCF;
-use wcf\util\ClassUtil;
 
 /**
  * Represents a sitemap entry.
@@ -49,7 +48,7 @@ class Sitemap extends DatabaseObject {
 				throw new SystemException("Unable to find class '".$this->className."' for sitemap '".$this->sitemapName."'");
 			}
 			
-			if (!ClassUtil::isInstanceOf($this->className, 'wcf\system\sitemap\ISitemapProvider')) {
+			if (!is_subclass_of($this->className, 'wcf\system\sitemap\ISitemapProvider')) {
 				throw new SystemException("'".$this->className."' does not implement 'wcf\system\sitemap\ISitemapProvider'");
 			}